赛程管理 start code

+++ code/src/main/java/cn/edu/tsinghua/thubp/ThubpApplication.java
@@ -7,6 +7,7 @@ import cn.edu.tsinghua.thubp.user.repository.UserRepository;
 import cn.edu.tsinghua.thubp.web.service.SequenceGeneratorService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -29,33 +30,39 @@ public class ThubpApplication implements CommandLineRunner {
     @Autowired
     private MongoTemplate mongoTemplate;
 
+    @Value("${spring.profiles.active}")
+    private String profile;
+
     public static void main(String[] args) {
         SpringApplication.run(ThubpApplication.class, args);
     }
 
     @Override
     public void run(String... args) {
-        // 清空数据库
-        log.info("清空数据库");
-        for (String name: mongoTemplate.getCollectionNames()) {
-            mongoTemplate.dropCollection(name);
-        }
-        // 初始化一个 admin 用户
-        log.info("创建 root 用户");
-        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
-        if (!userRepository.existsByThuId("2018000000")) {
-            User user = User.builder()
-                    .userId(sequenceGeneratorService.generateSequence(User.SEQUENCE_NAME))
-                    .gender(Gender.UNKNOWN)
-                    .enabled(true)
-                    .username("root")
-                    .password(bCryptPasswordEncoder.encode("root"))
-                    .role(RoleType.ROOT)
-                    .mobile("10000000000")
-                    .email("thubp@tsinghua.edu.cn")
-                    .thuId("2018000000")
-                    .build();
-            userRepository.save(user);
+        // dev 环境下运行命令
+        if (profile.equals("dev")) {
+            // 清空数据库
+            log.info("清空数据库");
+            for (String name: mongoTemplate.getCollectionNames()) {
+                mongoTemplate.dropCollection(name);
+            }
+            // 初始化一个 admin 用户
+            log.info("创建 root 用户");
+            BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
+            if (!userRepository.existsByThuId("2018000000")) {
+                User user = User.builder()
+                        .userId(sequenceGeneratorService.generateSequence(User.SEQUENCE_NAME))
+                        .gender(Gender.UNKNOWN)
+                        .enabled(true)
+                        .username("root")
+                        .password(bCryptPasswordEncoder.encode("root"))
+                        .role(RoleType.ROOT)
+                        .mobile("10000000000")
+                        .email("thubp@tsinghua.edu.cn")
+                        .thuId("2018000000")
+                        .build();
+                userRepository.save(user);
+            }
         }
     }
 }
